refactor: restructure Agent 29 to match agentcore_template pattern
Aligns with the standardized template per repo owner feedback (#248).

Structure now follows agentcore_template/Agent 28 pattern:
- main.py at root (BedrockAgentCoreApp entrypoint)
- agent/agent_config/agent.py (agent creation + tools + task logic)
- agent/resources/ (data files)
- tests/test_agent.py
- pyproject.toml at root

Closes #251
