--- agents_catalog/29-Clinical-prior-auth/agentcore/.dockerignore
@@ -1,72 +1,0 @@
-# Build artifacts
-build/
-dist/
-*.egg-info/
-*.egg
-
-# Python cache
-__pycache__/
-__pycache__*
-*.py[cod]
-*$py.class
-*.so
-.Python
-
-# Virtual environments
-.venv/
-.env
-venv/
-env/
-ENV/
-
-# Testing
-.pytest_cache/
-.coverage
-.coverage*
-htmlcov/
-.tox/
-*.cover
-.hypothesis/
-.mypy_cache/
-.ruff_cache/
-
-# Development
-*.log
-*.bak
-*.swp
-*.swo
-*~
-.DS_Store
-
-# IDEs
-.vscode/
-.idea/
-
-# Version control
-.git/
-.gitignore
-.gitattributes
-
-# Documentation
-docs/
-
-# CI/CD
-.github/
-.gitlab-ci.yml
-.travis.yml
-
-# Project specific
-tests/
-
-# Bedrock AgentCore specific - keep config but exclude runtime files
-.bedrock_agentcore.yaml
-.dockerignore
-.bedrock_agentcore/
-
-# Keep wheelhouse for offline installations
-# wheelhouse/
-
-# Monorepo directories
-cdk/
-terraform/
-mcp/lambda/
