+++ agents_catalog/29-Clinical-prior-auth/agentcore/.gitignore
@@ -1,10 +1,10 @@
 __pycache__/
 *.pyc
 .venv/
-downloaded_files/
 *.egg-info/
 dist/
 build/
 .bedrock_agentcore/
 .bedrock_agentcore.yaml
 .dockerignore
+/tmp/
