+++ agents_catalog/29-Clinical-prior-auth/agentcore/pyproject.toml
@@ -1,15 +1,14 @@
-[build-system]
-requires = ["setuptools>=68", "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
-name = "clinical-prior-auth"
+name = "29-clinical-prior-auth"
 version = "0.1.0"
+description = "Clinical Prior Authorization Agent - AgentCore"
+readme = "README.md"
 requires-python = ">=3.12"
-
 dependencies = [
     "bedrock-agentcore >= 1.0.3",
+    "bedrock-agentcore-starter-toolkit >= 0.1.0",
     "boto3 >= 1.38.0",
+    "botocore >= 1.38.0",
     "pandas >= 1.5.0",
     "PyPDF2 >= 3.0.0",
     "requests >= 2.28.0",
@@ -20,10 +19,6 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
     "pytest >= 7.0.0",
-    "pytest-asyncio >= 0.21.0",
     "ruff >= 0.12.0",
     "bandit >= 1.9.0",
 ]
-
-[tool.setuptools.package-data]
-"tools" = ["data/*.json"]
