--- agents_catalog/29-Clinical-prior-auth/agentcore/src/model/load.py
@@ -1,8 +1,0 @@
-from strands.models import BedrockModel
-
-MODEL_ID = "us.anthropic.claude-sonnet-4-5-20250929-v1:0"
-
-
-def load_model() -> BedrockModel:
-    """Get Bedrock model client for prior auth agent."""
-    return BedrockModel(model_id=MODEL_ID)
